# Experience-Hub/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Experiment {
  /** Unique identifier (filename without extension) */
  id: string;
  title: string;
  description: string;
  language: string;
  code: string;
  /** Output image filename or text */
  output?: string | null;
  subject: string;
}

export interface SubjectGroup {
  subject: string;
  experiments: Experiment[];
}

export interface ExperimentsResponse {
  subjects: string[];
  experiments: Experiment[];
}

export interface ErrorResponse {
  error: string;
}