# Experience-Hub/lib/api-zod/src/generated/types/experiment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Experiment {
  /** Unique identifier (filename without extension) */
  id: string;
  title: string;
  description: string;
  language: string;
  code: string;
  /** Output image filename or text */
  output?: string | null;
  subject: string;
}