# Experience-Hub/lib/api-zod/src/generated/types/experimentsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Experiment } from "./experiment";

export interface ExperimentsResponse {
  subjects: string[];
  experiments: Experiment[];
}